requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.6.0
pillow>=9.0.0
numpy>=1.23.0
streamlit>=1.28.0
torch>=1.8.0
torchvision>=0.9.0
pyyaml>=6.0

